fix: Rename click-handler to toggle-handler
BREAKING CHANGE: The click handler was actually being called whenever a toggle
element was clicked. It was also possible to click on places in a tree node that
wouldn't result in the callback firing... Needless to say this was a poorly
named attribute.

To migrate follow the example below:

Before:

<div ivh-treeview="data"
    ivh-treeview-click-handler="myHandler"></div>

ivhTreeviewOptionsProvider.set({
  clickHandler: myHandler
});

After:

<div ivh-treeview="data"
    ivh-treeview-toggle-handler="myHandler"></div>

ivhTreeviewOptionsProvider.set({
  toggleHandler: myHandler
});

Closes #45