feature: Allow for checkboxes to not propagate state to children (#201)
* feature: Allow for checkboxes to not propagate state to children

There is new option `disableCheckboxSelectionPropagation` that id set to `true` disables propagation of state of checkboxes to children and switches off validation of parents state. Thus the checkboxes are independent on each other and selection of checkbox on any level is relevant only to it's item.

Closes #195

* feature: Allow for checkboxes to not propagate state to children

There is new option `disableCheckboxSelectionPropagation` that id set to `true` disables propagation of state of checkboxes to children and switches off validation of parents state. Thus the checkboxes are independent on each other and selection of checkbox on any level is relevant only to it's item.

Closes #195

* Add built files

* Add test of disableCheckboxSelectionPropagation